requirements.txt
requests
bs4
django-cors-headers

